warning fixes for alpha-linux
2005-03-25  Bryan Ãstergaard  <kloeri@gentoo.org>

Warning fixes for alpha-linux.

* kaffe/kaffeh/kaffeh-support.h,
kaffe/kaffeh/support.c: (addCode, addLineNumbers,
addLocalVariables, addCheckedExceptions): Use size_t
instead of uint32 for len parameter.

* kaffe/kaffevm/classMethod.c,
kaffe/kaffevm/classMethod.h: (addInnerClasses) Use size_t
instead of uint32 for len parameter.

* kaffe/kaffevm/code.c,
kaffe/kaffevm/code.h: (addCode, addLineNumbers,
addLocalVariables, addCheckedExceptions) Use size_t
instead of uint32 for len parameter.

2005-03-25  Dalibor Topic  <robilad@kaffe.org>

* kaffe/kaffevm/readClass.c (readAttributes):
Cast len to size_t to avoid new warnings on 64 bit platforms.

+++ kaffe/kaffeh/kaffeh-support.h
@@ -28,10 +28,10 @@ extern void endJniInclude(void);
 extern void kaffeh_findClass(const char *nm);
 
 /* support.c */
-extern bool addCode(Method* m, uint32 len, classFile* fp, errorInfo *einfo);	
-extern bool addLineNumbers(Method* m, uint32 len, classFile* fp, errorInfo *info);
-extern bool addLocalVariables(Method* m, uint32 len, classFile* fp, errorInfo *info);
-extern bool addCheckedExceptions(Method* m, uint32 len, classFile* fp, errorInfo *info);
+extern bool addCode(Method* m, size_t len, classFile* fp, errorInfo *einfo);	
+extern bool addLineNumbers(Method* m, size_t len, classFile* fp, errorInfo *info);
+extern bool addLocalVariables(Method* m, size_t len, classFile* fp, errorInfo *info);
+extern bool addCheckedExceptions(Method* m, size_t len, classFile* fp, errorInfo *info);
 
 /* sig.c */
 extern const char* translateSig(const char*, const char**, int*);
